@@ -49,57 +50,56 @@
     "@babel/preset-env": "^7.8.3",
     "@babel/preset-react": "^7.8.3",
     "@babel/preset-typescript": "^7.10.4",
-    "@rollup/plugin-commonjs": "^11.0.1",
-    "@rollup/plugin-node-resolve": "^7.0.0",
-    "@rollup/plugin-replace": "^2.3.2",
-    "@rollup/plugin-typescript": "^5.0.1",
-    "@storybook/addon-actions": "^5.3.19",
-    "@storybook/addon-docs": "^5.3.19",
-    "@storybook/addon-links": "^5.3.19",
-    "@storybook/addons": "^5.3.19",
-    "@storybook/preset-scss": "^1.0.2",
-    "@storybook/react": "^5.3.19",
-    "@storybook/storybook-deployer": "^2.8.6",
-    "@svgr/rollup": "^5.1.0",
+    "@rollup/plugin-babel": "^5.3.0",
+    "@rollup/plugin-commonjs": "^19.0.0",
+    "@rollup/plugin-eslint": "^8.0.1",
+    "@rollup/plugin-node-resolve": "^7.1.3",
+    "@rollup/plugin-replace": "^2.4.2",
+    "@rollup/plugin-typescript": "^8.2.1",
+    "@storybook/addon-actions": "^6.3.4",
+    "@storybook/addon-docs": "^6.3.4",
+    "@storybook/addon-links": "^6.3.4",
+    "@storybook/addons": "^6.3.4",
+    "@storybook/preset-scss": "^1.0.3",
+    "@storybook/react": "^6.3.4",
+    "@storybook/storybook-deployer": "^2.8.10",
+    "@svgr/rollup": "^5.5.0",
     "@svgr/webpack": "^5.1.0",
     "@typescript-eslint/eslint-plugin": "^3.6.0",
     "@typescript-eslint/parser": "^3.6.0",
+    "@wojtekmaj/enzyme-adapter-react-17": "^0.6.2",
     "autoprefixer": "^9.7.4",
     "babel-eslint": "^10.0.3",
     "babel-jest": "^25.1.0",
     "babel-loader": "^8.0.6",
-    "babel-preset-es2015-rollup": "^3.0.0",
     "browserslist": "^4.14.5",
     "caniuse-lite": "^1.0.30001148",
     "css-loader": "^3.4.2",
     "enzyme": "^3.11.0",
-    "enzyme-adapter-react-16": "^1.15.2",
     "eslint": "^6.8.0",
     "eslint-config-airbnb": "^18.0.1",
     "eslint-plugin-import": "^2.20.0",
     "eslint-plugin-jsx-a11y": "^6.2.3",
     "eslint-plugin-react": "^7.18.0",
     "eslint-plugin-react-hooks": "^1.7.0",
-    "jest": "^25.1.0",
+    "jest": "^27.0.6",
     "jsdom": "^16.1.0",
+    "node-sass": "^6.0.1",
     "plop": "^2.5.3",
-    "react-test-renderer": "^16.8.6",
-    "rollup": "^1.30.1",
-    "rollup-plugin-babel": "^4.3.3",
-    "rollup-plugin-copy": "^3.3.0",
-    "rollup-plugin-eslint": "^7.0.0",
-    "rollup-plugin-extensions": "^0.1.0",
-    "rollup-plugin-peer-deps-external": "^2.2.0",
-    "rollup-plugin-postcss": "^2.0.6",
-    "rollup-plugin-scss": "^2.1.0",
+    "postcss": "^8.3.5",
+    "react-test-renderer": "^17.0.2",
+    "rollup": "^2.53.1",
+    "rollup-plugin-copy": "^3.4.0",
+    "rollup-plugin-postcss": "^4.0.0",
+    "rollup-plugin-scss": "^3.0.0",
     "rollup-plugin-size-snapshot": "^0.12.0",
-    "sass-loader": "^8.0.2",
+    "sass-loader": "^10.2.0",
     "style-loader": "^1.1.3",
     "stylelint": "^13.0.0",
     "stylelint-config-sass-guidelines": "^7.0.0",
     "ts-loader": "^7.0.5",
     "tslib": "^2.0.0",
-    "typescript": "^3.9.6"
+    "typescript": "^4.3.5"
   },
   "browserslist": [
     "defaults",
