@@ -1,9 +1,9 @@
 // plugins
-import babel from 'rollup-plugin-babel';
-import external from 'rollup-plugin-peer-deps-external';
+import { babel } from '@rollup/plugin-babel';
+// import external from 'rollup-plugin-peer-deps-external';
 import commonjs from '@rollup/plugin-commonjs';
 import resolve from '@rollup/plugin-node-resolve';
-import extensions from 'rollup-plugin-extensions';
+// import extensions from 'rollup-plugin-extensions';
 import svgr from '@svgr/rollup';
 import scss from 'rollup-plugin-scss';
 import postcss from 'rollup-plugin-postcss';
@@ -24,13 +24,13 @@ module.exports = ({
   // To bundle split
   input: {
     index: 'src/index.js',
-    App: 'src/smart-components/App',
-    ChannelList: 'src/smart-components/ChannelList',
-    ChannelSettings: 'src/smart-components/ChannelSettings',
-    Channel: 'src/smart-components/Conversation',
-    OpenChannel: 'src/smart-components/OpenchannelConversation',
-    MessageSearch: 'src/smart-components/MessageSearch',
-    OpenChannelSettings: 'src/smart-components/OpenChannelSettings',
+    App: 'src/smart-components/App/index.jsx',
+    ChannelList: 'src/smart-components/ChannelList/index.jsx',
+    ChannelSettings: 'src/smart-components/ChannelSettings/index.jsx',
+    Channel: 'src/smart-components/Conversation/index.jsx',
+    OpenChannel: 'src/smart-components/OpenchannelConversation/index.tsx',
+    MessageSearch: 'src/smart-components/MessageSearch/index.tsx',
+    OpenChannelSettings: 'src/smart-components/OpenChannelSettings/index.tsx',
     SendbirdProvider: 'src/lib/Sendbird.jsx',
   },
   output: [
@@ -45,7 +45,14 @@ module.exports = ({
       sourcemap: true,
     },
   ],
-  // external: ['sendbird', 'prop-types', 'react', 'react-dom'],
+  external: [
+    'sendbird',
+    'prop-types',
+    'react',
+    'react-dom',
+    'css-vars-ponyfill',
+    'date-fns',
+  ],
   plugins: [
     postcss({
       preprocessor: (content, id) => new Promise((resolvecss) => {
@@ -60,20 +67,21 @@ module.exports = ({
       extensions: ['.sass', '.scss', '.css'],
     }),
     replace({
+      preventAssignment: false,
       exclude: 'node_modules/**',
       [APP_VERSION_STRING]: pkg.version,
       [IS_ROLLUP]: IS_ROLLUP_REPLACE,
     }),
     typescript({ jsx: 'preserve' }),
-    external(),
-    extensions({
-      // Supporting Typescript files
-      // Uses ".mjs, .js" by default
-      extensions: ['.tsx', '.ts', '.jsx', '.js'],
-      // Resolves index dir files based on supplied extensions
-      // This is enable by default
-      resolveIndex: true,
-    }),
+    // external(),
+    // extensions({
+    //   // Supporting Typescript files
+    //   // Uses ".mjs, .js" by default
+    //   extensions: ['.tsx', '.ts', '.jsx', '.js'],
+    //   // Resolves index dir files based on supplied extensions
+    //   // This is enable by default
+    //   resolveIndex: true,
+    // }),
     svgr(),
     babel({
       presets: [
