@@ -18,8 +18,10 @@ export default function ImageRenderer({
   height,
   defaultComponent,
   circle,
+  placeHolder, // a function returing JSX / (style) => Element
 }) {
   const [showDefaultComponent, setShowDefaultComponent] = useState(false);
+  const [showPlaceHolder, setShowPlaceHolder] = useState(true);
 
   const DefaultComponent = useMemo(() => {
     if (typeof defaultComponent === 'function') {
@@ -28,22 +30,55 @@ export default function ImageRenderer({
     return defaultComponent;
   }, [defaultComponent]);
 
+  const PlaceHolder = useMemo(() => {
+    if (placeHolder && typeof placeHolder === 'function') {
+      return placeHolder({
+        style: {
+          width,
+          height,
+          position: 'absolute',
+          display: 'flex',
+          justifyContent: 'center',
+          alignItems: 'center',
+        },
+      });
+    }
+    return null;
+  }, [placeHolder]);
+
+  const HiddenImageLoader = useMemo(() => {
+    setShowDefaultComponent(false);
+    // reset the state when url is changed
+    return (
+      <img
+        className="sendbird-image-renderer__hidden-image-loader"
+        src={url}
+        alt={alt}
+        onLoad={() => setShowPlaceHolder(false)}
+        onError={() => setShowDefaultComponent(true)}
+      />
+    );
+  }, [url]);
+
   return (
     <div
       className={[
         ...(Array.isArray(className) ? className : [className]),
         'sendbird-image-renderer',
       ].join(' ')}
+      style={{ width, height }}
     >
+      {showPlaceHolder && PlaceHolder}
       {
         showDefaultComponent
-          ? (DefaultComponent)
+          ? DefaultComponent
           : (
             <div
               className="sendbird-image-renderer__image"
               style={{
                 width,
                 height,
+                position: 'absolute',
                 backgroundRepeat: 'no-repeat',
                 backgroundPosition: 'center',
                 backgroundSize: 'cover',
@@ -53,12 +88,7 @@ export default function ImageRenderer({
             />
           )
       }
-      <img
-        className="sendbird-image-renderer__hidden-image-loader"
-        alt={alt}
-        onError={() => setShowDefaultComponent(true)}
-        src={url}
-      />
+      {HiddenImageLoader}
     </div>
   );
 }
@@ -82,13 +112,15 @@ ImageRenderer.propTypes = {
     PropTypes.element,
     PropTypes.func,
   ]),
+  placeHolder: PropTypes.func,
   circle: PropTypes.bool,
 };
 ImageRenderer.defaultProps = {
   className: '',
+  defaultComponent: null,
+  placeHolder: null,
   alt: '',
   width: null,
   height: null,
-  defaultComponent: null,
   circle: false,
 };
